Update domain testing code to handle new drawing method
Changed the scope of a few classes to allow for access within the renderer.
Position is now generic to allow for the conversion between ints and doubles,
however this may change in the future due to typing choices by domain.
Animator now only specifies a single type of animation, because static animations
can be represented using dynamic animations.