Implement rendering classes
These classes are to handle the rendering of the game.
Currently, they are unbound from each other as they are incomplete.
Animation and AnimationHandler handle the sequencing and implementation of animated entities.
Drawable acts as an override safety check, however it might be removed at a later point.
Renderer handles how elements are drawn to the panel, specifically handling screenspace culling,
and conversion from domain coordinates to screen coordinates.