Update RenderPanel to handle tilemap and entity drawing
RenderPanel now has tick and draw methods to handle the ticking and drawing of specific tiles
and entities within the scene. RenderPanel also implements a two step culling routine to ensure that
textures are culled correctly if they cannot be seen on screen. Currently, this just shows a simple
texture moving over a tile plane as a proof of concept, however when connected to domain, this
provides a powerful rendering system.

In future, aspects of the renderpanel may be delegated to subclasses such as AnimationHandler,
or to class methods, however they are quite functional right now as lambda variables.